add ability to jump to first/last window
When changing to a tab we should jump to the most
natural window-location. This means that when
we move left (h) we should land on the right side
of the resulting tab and when we move right we should
land on the left side of the resulting tab.

When jumping to the first tab, we will
jump to the left side and on the last tab
we will jump to the right side (hoping
that this is a sensible decision).